Extract waveform visual bands
